Update php-mailer-smtp.php
In this code, we use the PHPMailer library to send an email. We first require the necessary PHPMailer classes and then create a new instance of the PHPMailer class. We set various properties of the $mail object, such as the email host, port, and authentication credentials. We then set the "From" email address, recipient email address, subject, and body of the email. Finally, we use the Send() method to send the email and display an error message if it fails to send.

Download library from here https://github.com/PHPMailer/PHPMailer