Orange SMS sending
In this code, we use the Orange SMS Pro API to send an SMS message. We set various properties of the SMS message, such as the recipient's phone number, subject, signature, and body. We then concatenate these message data and generate a hash-based message authentication code (HMAC) with the SHA-1 algorithm. We create an array of SMS data and encode it into a format that can be sent via HTTP POST. Finally, we use the cURL library to send the SMS data to the API endpoint, using basic HTTP authentication with the HMAC key and API token.